remove ugly mix
