+++ src/MonthCalendar.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
+import classnames from 'classnames';
 import KeyCode from 'rc-util/lib/KeyCode';
 import { polyfill } from 'react-lifecycles-compat';
 import moment, { Moment } from 'moment';
 import CalendarHeader from './calendar/CalendarHeader';
 import CalendarFooter from './calendar/CalendarFooter';
-import { calendarMixinWrapper, calendarMixinDefaultProps } from './mixin/CalendarMixin';
-import { commonMixinWrapper, defaultProp } from './mixin/CommonMixin';
+import { calendarMixinDefaultProps, getNowByCurrentStateValue } from './mixin/CalendarMixin';
+import { defaultProp } from './mixin/CommonMixin';
 import { CalendarTypeMode } from './date/DateInput';
 
 interface MonthCalendarProps {
@@ -19,10 +20,14 @@ interface MonthCalendarProps {
   renderFooter?: (mode: CalendarTypeMode) => void;
   monthCellContentRender?: () => React.ReactNode;
   locale: { [key: string]: any };
+  onChange?: (value: Moment) => void;
+  className?: string;
+  style?: React.CSSProperties;
+  onSelect?: (value: Moment, cause?) => void;
 }
 
 interface MonthCalendarState {
-  mode: CalendarTypeMode;
+  mode?: CalendarTypeMode;
   value?: Moment;
   selectedValue?: Moment;
 }
@@ -36,12 +41,45 @@ class MonthCalendar extends React.Component<MonthCalendarProps, MonthCalendarSta
     this.state = {
       mode: 'month',
       value: props.value || props.defaultValue || moment(),
+      selectedValue: props.value || props.defaultValue || moment(),
     };
   }
 
-  onSelect: (stateValue: Moment) => React.ReactNode;
+  static getDerivedStateFromProps(nextProps: MonthCalendarProps, prevState: MonthCalendarState) {
+    const { value, selectedValue } = nextProps;
+    const newState: MonthCalendarState = {};
 
-  setValue: (stateValue: Moment) => React.ReactNode;
+    if ('value' in nextProps) {
+      newState.value =
+        value || nextProps.defaultValue || getNowByCurrentStateValue(prevState.value);
+    }
+    if ('selectedValue' in nextProps) {
+      newState.selectedValue = selectedValue;
+    }
+
+    return newState;
+  }
+
+  onSelect = (value, cause?) => {
+    if (value) {
+      this.setValue(value);
+    }
+    this.setSelectedValue(value, cause);
+  };
+
+  setSelectedValue = (selectedValue, cause) => {
+    // 一个重来没用过得值，但是有个测试覆盖到了它
+    // 这里强行用了一下
+    const { selectedValue: stateValue } = this.state;
+    if (!('selectedValue' in this.props) && stateValue) {
+      this.setState({
+        selectedValue,
+      });
+    }
+    if (this.props.onSelect) {
+      this.props.onSelect(selectedValue, cause);
+    }
+  };
 
   onKeyDown = event => {
     const { keyCode } = event;
@@ -97,7 +135,52 @@ class MonthCalendar extends React.Component<MonthCalendarProps, MonthCalendarSta
     }
   };
 
-  renderRoot: (option: { children: React.ReactNode; className: string }) => React.ReactNode;
+  setValue = (value: Moment) => {
+    const originalValue = this.state.value;
+    if (!('value' in this.props)) {
+      this.setState({
+        value,
+      });
+    }
+    if (
+      (originalValue && value && !originalValue.isSame(value)) ||
+      (!originalValue && value) ||
+      (originalValue && !value)
+    ) {
+      this.props.onChange(value);
+    }
+  };
+
+  rootInstance: HTMLElement;
+
+  saveRoot = root => {
+    this.rootInstance = root;
+  };
+
+  onBlur: (event: React.FocusEvent<HTMLDivElement>) => void;
+
+  renderRoot = newProps => {
+    const { props } = this;
+    const { prefixCls } = props;
+
+    const className = {
+      [prefixCls]: 1,
+      [props.className]: !!props.className,
+      [newProps.className]: !!newProps.className,
+    };
+    return (
+      <div
+        ref={this.saveRoot}
+        className={`${classnames(className)}`}
+        style={this.props.style}
+        tabIndex={0}
+        onKeyDown={this.onKeyDown}
+        onBlur={this.onBlur}
+      >
+        {newProps.children}
+      </div>
+    );
+  };
 
   render() {
     const { props, state } = this;
@@ -128,4 +211,4 @@ class MonthCalendar extends React.Component<MonthCalendarProps, MonthCalendarSta
   }
 }
 
-export default polyfill(calendarMixinWrapper(commonMixinWrapper(MonthCalendar)));
+export default polyfill(MonthCalendar);
