+++ src/mixin/CommonMixin.tsx
@@ -1,5 +1,4 @@
 import PropTypes from 'prop-types';
-import React from 'react';
 import enUs from '../locale/en_US';
 
 function noop() {}
@@ -31,49 +30,3 @@ export const defaultProp = {
     return null;
   },
 };
-
-export const commonMixinWrapper = (ComposeComponent: React.ComponentClass<any, any>) =>
-  class extends ComposeComponent {
-    static displayName = 'CommonMixinWrapper';
-
-    static defaultProps = ComposeComponent.defaultProps;
-
-    static getDerivedStateFromProps = ComposeComponent.getDerivedStateFromProps;
-
-    shouldComponentUpdate(nextProps) {
-      return this.props.visible || nextProps.visible;
-    }
-
-    getFormat = () => {
-      let { format } = this.props;
-      const { locale, timePicker } = this.props;
-      if (!format) {
-        if (timePicker) {
-          format = locale.dateTimeFormat;
-        } else {
-          format = locale.dateFormat;
-        }
-      }
-      return format;
-    };
-
-    focusElement: HTMLElement;
-
-    rootInstance: HTMLElement;
-
-    focus = () => {
-      if (this.focusElement) {
-        this.focusElement.focus();
-      } else if (this.rootInstance) {
-        this.rootInstance.focus();
-      }
-    };
-
-    saveFocusElement = (focusElement: HTMLElement) => {
-      this.focusElement = focusElement;
-    };
-
-    saveRoot = root => {
-      this.rootInstance = root;
-    };
-  };
